+++ sdk/storage/azure-storage-blob-nio/CHANGELOG.md
@@ -1,7 +1,7 @@
 # Release History
 
 ## 12.0.0-beta.2 (Unreleased)
-
+- Added checks to ensure file system has not been closed before operating on data
 
 ## 12.0.0-beta.1 (2020-07-17)
 - Initial Release. Please see the README for more information.